MAINT: Move parts of the `assetVersion` script to its own common file
